Fixed map issues. Added electron functionality. Added development electron script. Added electron package script
