+++ README.md
@@ -39,6 +39,14 @@ Instead, it will copy all the configuration files and the transitive dependencie
 
 You don’t have to ever use `eject`. The curated feature set is suitable for small and middle deployments, and you shouldn’t feel obligated to use this feature. However we understand that this tool wouldn’t be useful if you couldn’t customize it when you are ready for it.
 
+### `yarn dist`
+
+Package application into windows application. Don't forget to build application before run this comand
+
+### `yarn electron-dev`
+
+Run electron application in development mode. It means that application will running on [http://localhost:3000](http://localhost:3000).
+
 ## Learn More
 
 You can learn more in the [Create React App documentation](https://facebook.github.io/create-react-app/docs/getting-started).
