+++ package.json
@@ -12,8 +12,10 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.21.1",
     "dotenv": "^8.2.0",
+    "electron-is-dev": "^2.0.0",
     "eslint-plugin-prettier": "^3.3.1",
     "firebase": "^8.2.9",
+    "is-electron": "^2.2.0",
     "leaflet": "^1.7.1",
     "moment": "^2.29.1",
     "react": "^17.0.1",
@@ -45,13 +47,15 @@
   },
   "scripts": {
     "start": "react-app-rewired start",
-    "electron-start": "electron .",
+    "electron-dev": "concurrently \"SET BROWSER=none&&npm start\" \"wait-on http://localhost:3000 && electron .\"",
     "build": "react-app-rewired build",
     "test": "react-app-rewired test",
     "eject": "react-scripts eject",
     "deploy": "firebase deploy",
     "login": "firebase login",
-    "init": "firebase init"
+    "init": "firebase init",
+    "pack": "electron-builder --dir",
+    "dist": "electron-builder"
   },
   "eslintConfig": {
     "extends": [
@@ -71,12 +75,28 @@
       "last 1 safari version"
     ]
   },
+  "build": {
+    "appId": "com.example.currency-converter",
+		"files": [
+      "**/*",
+      "*",
+      "**",
+			"/build/**/*"
+    ],
+    "directories": {
+      "buildResources": "assets"
+    }
+  },
   "description": "idle currency converter project",
-  "main": "index.js",
+  "main": "public/electron.js",
   "repository": "https://github.com/stanislavlebedyantsev/currency-converter-with-PWA.git",
   "author": "stanislavlebedyantsev <stanislav.lebedyantsev@gmail.com>",
   "license": "MIT",
   "devDependencies": {
-    "electron": "^12.0.1"
-  }
+    "concurrently": "^6.0.0",
+    "electron": "^12.0.1",
+    "electron-builder": "^22.10.5",
+    "wait-on": "^5.2.1"
+  },
+  "homepage": "./"
 }
